introduce WebSocket views with support for signals through Redis PubSub
This commit introduce the WSView class (WebSocket view), which is for
handling WebSocket connection through the following:

  - WebSocket signals, through the Redis' PubSub utility
  - processing WebSocket messages from a client
  - checking whether a user has access to the view

Each app may subclass WSView to write its own WebSocket views, which
can then be used in WebSocket URLconf as a view (The entirety of the
class should be passed as the view 'func', as contrary to Django's
normal views. This is basically an (ab)use of Django's URLconf system).

There are resemblances between Django's class-based views and WebSocket
views, but they are quite different, and they are not meant to be used
the same way. For more information, see the docstring and comments in
the django_project/websocket/views.py.