Add the simulator machinery.
The simulator was something that had to be done as a whole,
and different parts of it were so interconnected that I could
not see an appropriate occasion to commit until the whole
thing was basically done (except for test cases, even though
the test facilities are implemented).

I've tried to explain different parts of the code through
comprehensive comments. If tweaking was necessary, it's
advised to check the comments first, especially their
warnings.