bugfix: use cursor for each raw SQL execution
Storing a single cursor would make the program unusable in case there
was a disconnection. This would happen when we stop the docker containers
without removing them, and trying to restart them again (which would
try to use the previous state of the container, but it wouldn't work
since the database connection is lost [not totally sure]).